Unify grappler optimizer and rewrite-pass
* Remove duplicated code in ngraph-optimizer and make ngraph-optimizer use
 rewrite-pass.
* Remove unnecessary NgraphEncapsulationPass
* Move ngraph-optimizer.{cc,h} to ngraph_bridge/
