+++ ngraph_bridge/ngraph_rewrite_pass.cc
@@ -28,28 +28,13 @@
 #include "encapsulate_clusters.h"
 #include "log.h"
 #include "mark_for_clustering.h"
+#include "ngraph_rewrite_pass.h"
 #include "utils.h"
 
 using namespace std;
 
 namespace tensorflow {
 namespace ngraph_bridge {
-
-class NGraphRewritePass : public GraphOptimizationPass {
- public:
-  virtual Status Run(const GraphOptimizationPassOptions& options) = 0;
-
- protected:
-  // Returns a fresh "serial number" to avoid filename collisions in the graph
-  // dumps.
-  static int FreshIndex() {
-    mutex_lock l(s_serial_counter_mutex);
-    return s_serial_counter++;
-  }
-
-  static int s_serial_counter GUARDED_BY(s_serial_counter_mutex);
-  static mutex s_serial_counter_mutex;
-};
 
 int NGraphRewritePass::s_serial_counter = 0;
 mutex NGraphRewritePass::s_serial_counter_mutex;
@@ -64,70 +49,71 @@ mutex NGraphRewritePass::s_serial_counter_mutex;
 //   3. Cluster Deassignment [deassign_clusters.cc]
 //   4. Cluster Encapsulation [encapsulate_clusters.cc]
 
-class NGraphEncapsulationPass : public NGraphRewritePass {
- public:
-  Status Run(const GraphOptimizationPassOptions& options) override {
-    // If we don't get a main graph, log that fact and bail.
-    if (options.graph == nullptr) {
-      NGRAPH_VLOG(0) << "NGraphEncapsulationPass: options.graph == nullptr";
-      return Status::OK();
-    }
+Status NGraphRewritePass::Rewrite(
+    Graph* graph, std::set<string> skip_these_nodes,
+    std::unordered_map<std::string, std::string> config_map) {
+  int idx = FreshIndex();
+  // If requested, dump unmarked graphs.
+  util::DumpTFGraph(graph, idx, "unmarked");
 
-    // For filename generation purposes, grab a fresh index. This is just an
-    // arbitrary integer to avoid filename collisions resulting from subsequent
-    // runs of this pass.
-    int idx = FreshIndex();
-
-    tensorflow::Graph* graph = options.graph->get();
-    // If requested, dump unmarked graphs.
-    util::DumpTFGraph(graph, idx, "unmarked");
-
-    // If ngraph is disabled via ngraph_bridge api or NGRAPH_TF_DISABLE is set
-    // we will not do anything; all subsequent
-    // passes become a no-op.
-    bool ngraph_not_enabled =
-        (!api::IsEnabled()) || (std::getenv("NGRAPH_TF_DISABLE") != nullptr);
-    bool already_processed = util::IsAlreadyProcessed(graph);
-    if (!already_processed && ngraph_not_enabled) {
-      NGRAPH_VLOG(0) << "NGraph is available but disabled.";
-    }
-    if (ngraph_not_enabled || already_processed) {
-      NGRAPH_VLOG(1) << std::string("Rewrite pass will not run because ") +
-                            (already_processed ? "graph is already preprocessed"
-                                               : "ngraph is disabled");
-      ClusterManager::EvictAllClusters();
-      return Status::OK();
-    }
-
-    // Now Process the Graph
-
-    // 1. Mark for clustering then, if requested, dump the graphs.
-    std::set<string> skip_these_nodes = {};
-    TF_RETURN_IF_ERROR(MarkForClustering(graph, skip_these_nodes));
-    util::DumpTFGraph(graph, idx, "marked");
-
-    // 2. Assign clusters then, if requested, dump the graphs.
-    TF_RETURN_IF_ERROR(AssignClusters(graph));
-    util::DumpTFGraph(graph, idx, "clustered");
-
-    // 3. Deassign trivial clusters then, if requested, dump the graphs.
-    TF_RETURN_IF_ERROR(DeassignClusters(graph));
-    util::DumpTFGraph(graph, idx, "declustered");
-
-    // 4. Encapsulate clusters then, if requested, dump the graphs.
-    std::unordered_map<std::string, std::string> config_map;
-    auto status = EncapsulateClusters(graph, idx, config_map);
-    if (status != Status::OK()) {
-      return status;
-    }
-
-    util::DumpTFGraph(graph, idx, "encapsulated");
+  // If ngraph is disabled via ngraph_bridge api or NGRAPH_TF_DISABLE is set
+  // we will not do anything; all subsequent
+  // passes become a no-op.
+  bool ngraph_not_enabled =
+      (!api::IsEnabled()) || (std::getenv("NGRAPH_TF_DISABLE") != nullptr);
+  bool already_processed = util::IsAlreadyProcessed(graph);
+  if (!already_processed && ngraph_not_enabled) {
+    NGRAPH_VLOG(0) << "NGraph is available but disabled.";
+  }
+  if (ngraph_not_enabled || already_processed) {
+    NGRAPH_VLOG(1) << std::string("Rewrite pass will not run because ") +
+                          (already_processed ? "graph is already preprocessed"
+                                             : "ngraph is disabled");
+    ClusterManager::EvictAllClusters();
     return Status::OK();
   }
-};
+
+  // Now Process the Graph
+
+  // 1. Mark for clustering then, if requested, dump the graphs.
+  TF_RETURN_IF_ERROR(MarkForClustering(graph, skip_these_nodes));
+  util::DumpTFGraph(graph, idx, "marked");
+
+  // 2. Assign clusters then, if requested, dump the graphs.
+  TF_RETURN_IF_ERROR(AssignClusters(graph));
+  util::DumpTFGraph(graph, idx, "clustered");
+
+  // 3. Deassign trivial clusters then, if requested, dump the graphs.
+  TF_RETURN_IF_ERROR(DeassignClusters(graph));
+  util::DumpTFGraph(graph, idx, "declustered");
+
+  // 4. Encapsulate clusters then, if requested, dump the graphs.
+  auto status = EncapsulateClusters(graph, idx, config_map);
+  if (status != Status::OK()) {
+    return status;
+  }
+
+  util::DumpTFGraph(graph, idx, "encapsulated");
+  return Status::OK();
+}
+
+Status NGraphRewritePass::Run(const GraphOptimizationPassOptions& options) {
+  // If we don't get a main graph, log that fact and bail.
+  if (options.graph == nullptr) {
+    NGRAPH_VLOG(0) << "NGraphRewritePass: options.graph == nullptr";
+    return Status::OK();
+  }
+
+  // For filename generation purposes, grab a fresh index. This is just an
+  // arbitrary integer to avoid filename collisions resulting from subsequent
+  // runs of this pass.
+  return Rewrite(options.graph->get());
+}
 
 }  // namespace ngraph_bridge
 
+#ifndef NGRAPH_TF_USE_GRAPPLER_OPTIMIZER
 REGISTER_OPTIMIZATION(OptimizationPassRegistry::POST_REWRITE_FOR_EXEC, 0,
-                      ngraph_bridge::NGraphEncapsulationPass);
+                      ngraph_bridge::NGraphRewritePass);
+#endif
 }  // namespace tensorflow
